V0.2.1
Se cambia la altitud barometrica de flotante a entero, debido a que no aumentaba considerablemente la precisión y reduce la complejidad de la estructura.

+++ RFC_V1/Core/Inc/Libraries/BMP280.h
@@ -48,9 +48,8 @@ typedef struct BMP280_Data{
 	uint8_t 	ID;
 	int32_t		Temp;
 	uint32_t	Pressure;
-	float 		Barometric_Altitude;
+	uint16_t 	Barometric_Altitude;
 }BMP280_t;
-
 
 void BMP280_init(void);
 void BMP280_calculate(void);
